Refactor repair strategies
