--- src/main/java/sorald/SoraldAbstractRepair.java
@@ -1,131 +1,0 @@
-package sorald;
-
-import java.io.File;
-import java.lang.reflect.Constructor;
-import java.lang.reflect.InvocationTargetException;
-import java.util.Collection;
-import java.util.List;
-import java.util.function.Supplier;
-import sorald.processor.SoraldAbstractProcessor;
-import spoon.Launcher;
-import spoon.compiler.Environment;
-import spoon.processing.Processor;
-import spoon.reflect.CtModel;
-import spoon.reflect.declaration.CtElement;
-import spoon.reflect.reference.CtTypeReference;
-import spoon.reflect.visitor.DefaultImportComparator;
-import spoon.reflect.visitor.DefaultJavaPrettyPrinter;
-import spoon.reflect.visitor.ImportCleaner;
-import spoon.reflect.visitor.ImportConflictDetector;
-import spoon.reflect.visitor.PrettyPrinter;
-import spoon.support.JavaOutputProcessor;
-import spoon.support.sniper.SniperJavaPrettyPrinter;
-
-public abstract class SoraldAbstractRepair {
-    protected final GitPatchGenerator generator = new GitPatchGenerator();
-    protected SoraldConfig config;
-    protected int patchedFileCounter = 0;
-
-    public SoraldAbstractRepair(SoraldConfig config) {
-        this.config = config;
-        if (this.config.getGitRepoPath() != null) {
-            this.generator.setGitProjectRootDir(this.config.getGitRepoPath());
-        }
-    }
-
-    public abstract void repair();
-
-    protected void createPatches(String patchedFilePath, JavaOutputProcessor javaOutputProcessor) {
-        File patchDir = new File(this.config.getWorkspace() + File.separator + Constants.PATCHES);
-
-        if (!patchDir.exists()) {
-            patchDir.mkdirs();
-        }
-        List<File> list = javaOutputProcessor.getCreatedFiles();
-        if (!list.isEmpty()) {
-            String outputPath = list.get(list.size() - 1).getAbsolutePath();
-            generator.generate(
-                    patchedFilePath,
-                    outputPath,
-                    patchDir.getAbsolutePath()
-                            + File.separator
-                            + Constants.PATCH_FILE_PREFIX
-                            + this.patchedFileCounter);
-            this.patchedFileCounter++;
-        }
-    }
-
-    protected Launcher initLauncher(Launcher launcher, String outputDirPath) {
-        launcher.setSourceOutputDirectory(outputDirPath);
-        Environment env = launcher.getEnvironment();
-        env.setIgnoreDuplicateDeclarations(true);
-
-        // this is a workaround for https://github.com/INRIA/spoon/issues/3693
-        if (config.getPrettyPrintingStrategy() == PrettyPrintingStrategy.SNIPER) {
-            env.setPrettyPrinterCreator(() -> new SniperJavaPrettyPrinter(env));
-        }
-
-        // need to build the model before setting the pretty-printer as the preprocessors need
-        // data from the model
-        CtModel model = launcher.buildModel();
-
-        setPrettyPrinter(env, model);
-        return launcher;
-    }
-
-    private void setPrettyPrinter(Environment env, CtModel model) {
-        Supplier<? extends DefaultJavaPrettyPrinter> basePrinterCreator =
-                config.getPrettyPrintingStrategy() == PrettyPrintingStrategy.SNIPER
-                        ? createSniperPrinter(env)
-                        : createDefaultPrinter(env);
-        Supplier<PrettyPrinter> configuredPrinterCreator =
-                applyCommonPrinterOptions(basePrinterCreator, model);
-        env.setPrettyPrinterCreator(configuredPrinterCreator);
-    }
-
-    private static Supplier<PrettyPrinter> applyCommonPrinterOptions(
-            Supplier<? extends DefaultJavaPrettyPrinter> prettyPrinterCreator, CtModel model) {
-        Collection<CtTypeReference<?>> existingReferences = model.getElements(e -> true);
-        List<Processor<CtElement>> preprocessors =
-                List.of(
-                        new SelectiveForceImport(existingReferences),
-                        new ImportConflictDetector(),
-                        new ImportCleaner().setImportComparator(new DefaultImportComparator()));
-        return () -> {
-            DefaultJavaPrettyPrinter printer = prettyPrinterCreator.get();
-            printer.setIgnoreImplicit(false);
-            printer.setPreprocessors(preprocessors);
-            return printer;
-        };
-    }
-
-    private static Supplier<? extends DefaultJavaPrettyPrinter> createSniperPrinter(
-            Environment env) {
-        env.setCommentEnabled(true);
-        env.useTabulations(true);
-        env.setTabulationSize(4);
-        return () -> new SniperJavaPrettyPrinter(env);
-    }
-
-    private static Supplier<? extends DefaultJavaPrettyPrinter> createDefaultPrinter(
-            Environment env) {
-        return () -> new DefaultJavaPrettyPrinter(env);
-    }
-
-    protected SoraldAbstractProcessor createBaseProcessor(Integer ruleKey) {
-        try {
-            Class<?> processor = Processors.getProcessor(ruleKey);
-            if (processor != null) {
-                Constructor<?> cons = processor.getConstructor();
-                SoraldAbstractProcessor object = (SoraldAbstractProcessor) cons.newInstance();
-                return object;
-            }
-        } catch (InstantiationException
-                | IllegalAccessException
-                | InvocationTargetException
-                | NoSuchMethodException e) {
-            e.printStackTrace();
-        }
-        return null;
-    }
-}
